Add support for reading/writing GEODST CCCC files
GEODST is a geometry specification file defined on the CCCC standard
interface committee. This adds the ability to read and write these files
in ASCII and binary format.

This also moves NHFLUX support out from a __init__ into its own module
with no changes.
